Refactor CreditCardList styles, change from JS to TS
